requirements.txt
flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==8.4.0
pandas==2.2.2
python-dotenv==1.0.1